(REF) Extract base class `AbstractFlowsTest`
